# Optimise feature sampling via row-major (C) order feature array