@kieranricardo 2020-04-09
@huonw have you seen a speedup in the benchmark with the larger features?