@huonw 2020-04-09
> #1167: 3.2 s ± 136 ms per loop (mean ± std. dev. of 5 runs, 5 loops each)

I investigated this, and found that TF makes this new benchmark significantly faster... but only for large `SAMPLE_SIZE`:

![image](https://user-images.githubusercontent.com/1203825/78856805-0ad45b80-7a6b-11ea-927e-ee34e64e300d.png)

![image](https://user-images.githubusercontent.com/1203825/78856799-04de7a80-7a6b-11ea-947c-7cf36f0260d8.png)

These plots show the the times (in seconds) for the same configuration of the benchmark as in the table in the PR description (`feature_size = 1000, n_types = 1, type_arg = specify`), for different values of `SAMPLE_SIZE`. The current committed value of 50 in the microbenchmark is all the way on the left where NumPy has less overhead than Tensorflow, and the value used in the GraphSAGE benchmark (84.2k) is towards the right; after the transition point where Tensorflow is much faster than NumPy.

My theory about the transition point at 32k is some sort of cache effect, but it's unclear to me what.

<details><summary>Data</summary>

```python
import pandas as pd
data = pd.DataFrame(
    [
        (10, 0.0044, 0.217),
        (50, 0.015, 0.25),
        (100, 0.028, 0.367),
        (500, 0.61, 1.02),
        (1e3, 1.15, 1.78),
        (5e3, 7.6, 7.3),
        (10e3, 15, 14),
        (20e3, 28, 28),
        (30e3, 43, 42),
        (2**15, 46, 44),
        (33e3, 49, 44),
        (34e3, 83, 82),
        (35e3, 92, 86),
        (40e3, 103, 87),
        (50e3, 134, 91),
        (60e3, 156, 98),
        (2**16, 173, 105),
        (80e3, 210, 112),
        (2**17, 339, 160),
        (2**18, 692, 325),
        (2**19, 1400, 602)
    ],
    columns=["sample size", "numpy", "tensorflow"]
).set_index("sample size") / 1000
```

</details>

Also, I've opened #1228 to cover switching to using tensorflow for storing features.